refactor: panic module
